Accepting reviews

+++ Sources/ZMarkupParser/HTML/Processor/HTMLTagAttributeToMarkupStyleVisitor.swift
@@ -92,15 +92,15 @@ struct HTMLTagStyleAttributeToMarkupStyleVisitor: HTMLTagStyleAttributeVisitor {
     
     func visit(_ styleAttribute: LineHeightHTMLTagStyleAttribute) -> Result {
         guard let lineHeightFloat = self.convert(fromPX: value) else { return nil }
-        return MarkupStyle(paragraphStyle: MarkupStyleParagraphStyle(minimumLineHeight: CGFloat(lineHeightFloat), maximumLineHeight: CGFloat(lineHeightFloat)))
+        return MarkupStyle(paragraphStyle: MarkupStyleParagraphStyle(minimumLineHeight: lineHeightFloat, maximumLineHeight: lineHeightFloat))
     }
     
     func visit(_ styleAttribute: WordSpacingHTMLTagStyleAttribute) -> Result {
         guard let lineSpacing = self.convert(fromPX: value) else { return nil }
-        return MarkupStyle(paragraphStyle: MarkupStyleParagraphStyle(lineSpacing: CGFloat(lineSpacing)))
+        return MarkupStyle(paragraphStyle: MarkupStyleParagraphStyle(lineSpacing: lineSpacing))
     }
     
-    func convert(fromPX string: String) -> Int? {
+    func convert(fromPX string: String) -> CGFloat? {
         guard let regex = try? NSRegularExpression(pattern: "([0-9]+.?[0-9]*)p(x|t)"),
               let firstMatch = regex.firstMatch(in: string, options: [], range: NSRange(location: 0, length: string.utf16.count)),
               firstMatch.range(at: 1).location != NSNotFound,
@@ -108,7 +108,7 @@ struct HTMLTagStyleAttributeToMarkupStyleVisitor: HTMLTagStyleAttributeVisitor {
               let size = Float(String(string[range])) else {
             return nil
         }
-        return Int(size)
+        return CGFloat(size)
     }
 
     func visit(_ styleAttribute: TextAlignHTMLTagStyleAttribute) -> MarkupStyle? {
